refactor(taxes): extract OwnerBadge component
Move OwnerBadge to src/pages/taxes/components/OwnerBadge.tsx.
Taxes.tsx 1037→995L.

Part of #181
