+++ src/pages/taxes/Taxes.tsx
@@ -6,53 +6,11 @@ import type { TaxChecklistItem, TaxDocFile, TaxDocOwner, ChecklistCategory, TaxT
 import type { Account } from '../data/types'
 import { getStorageEstimate } from '../../utils/taxFileDB'
 import { fileToBase64, nextFileId } from './utils/fileHelpers'
+import OwnerBadge from './components/OwnerBadge'
 import '../../styles/Taxes.css'
 
 const CURRENT_YEAR = new Date().getFullYear()
 const MAX_FILE_SIZE = 10 * 1024 * 1024 // 10 MB
-
-/* ── owner badge ──────────────────────────────────────────── */
-const OwnerBadge: FC<{
-  owner: TaxDocOwner
-  primaryName: string
-  partnerName: string
-  primaryAvatar?: string
-  partnerAvatar?: string
-}> = ({ owner, primaryName, partnerName, primaryAvatar, partnerAvatar }) => {
-  const pi = (primaryName || 'P')[0].toUpperCase()
-  const si = (partnerName || 'S')[0].toUpperCase()
-  if (owner === 'joint') {
-    return (
-      <span className="tax-owner-group" title="Joint">
-        <span className="tax-owner-avatar tax-owner-primary">
-          {primaryAvatar ? <img src={primaryAvatar} alt="" /> : pi}
-        </span>
-        <span className="tax-owner-avatar tax-owner-partner">
-          {partnerAvatar ? <img src={partnerAvatar} alt="" /> : si}
-        </span>
-      </span>
-    )
-  }
-  const isPartner = owner === 'partner'
-  return (
-    <span
-      className={`tax-owner-avatar ${isPartner ? 'tax-owner-partner' : 'tax-owner-primary'}`}
-      title={isPartner ? partnerName : primaryName}
-    >
-      {isPartner ? (
-        partnerAvatar ? (
-          <img src={partnerAvatar} alt="" />
-        ) : (
-          si
-        )
-      ) : primaryAvatar ? (
-        <img src={primaryAvatar} alt="" />
-      ) : (
-        pi
-      )}
-    </span>
-  )
-}
 
 /* ── checklist item row ──────────────────────────────────── */
 const ChecklistRow: FC<{
